Consolidate AzureResourceGroup cloud test setup and group under ExternalCloudIntegration
Hoist the duplicated service-principal auth and resource-group provision/teardown
from the two cloud fixtures into a shared AzureResourceGroupCloudTestBase, and move
the fixtures into an ExternalCloudIntegration/ folder so the credential-free vs
live-Azure split is visible in the tree.
