--- source/Calamari.AzureResourceGroup.Tests/DeployAzureBicepTemplateCommandFixture.cs
@@ -1,128 +1,0 @@
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Azure;
-using Azure.Core;
-using Azure.ResourceManager;
-using Azure.ResourceManager.Resources;
-using Calamari.Azure;
-using Calamari.AzureResourceGroup.Bicep;
-using Calamari.CloudAccounts;
-using Calamari.Testing;
-using Calamari.Testing.Azure;
-using Calamari.Testing.Helpers;
-using Calamari.Testing.Requirements;
-using NUnit.Framework;
-
-namespace Calamari.AzureResourceGroup.Tests
-{
-    [TestFixture]
-    [Category(TestCategory.ExternalCloudIntegration)]
-    [WindowsTest] // NOTE: We should look at having the Azure CLI installed on Linux boxes so that these steps can be tested there, particularly if we're moving cloud to a Ubuntu Default Worker.
-    class DeployAzureBicepTemplateCommandFixture
-    {
-        string clientId;
-        string clientSecret;
-        string tenantId;
-        string subscriptionId;
-        string resourceGroupName;
-        string resourceGroupLocation;
-        ArmClient armClient;
-        static readonly CancellationTokenSource CancellationTokenSource = new CancellationTokenSource(TimeSpan.FromMinutes(5));
-        readonly CancellationToken cancellationToken = CancellationTokenSource.Token;
-        readonly string packagePath = TestEnvironment.GetTestPath("Packages", "Bicep");
-        SubscriptionResource subscriptionResource;
-
-        const string ParameterContent = """[{"Key":"storageAccountName","Value":"#{StorageAccountName}"},{"Key":"location","Value":"#{Location}"},{"Key":"sku","Value":"#{SKU}"}]""";
-
-        [OneTimeSetUp]
-        public async Task Setup()
-        {
-            var resourceManagementEndpointBaseUri =
-                Environment.GetEnvironmentVariable(AccountVariables.ResourceManagementEndPoint) ?? DefaultVariables.ResourceManagementEndpoint;
-            var activeDirectoryEndpointBaseUri =
-                Environment.GetEnvironmentVariable(AccountVariables.ActiveDirectoryEndPoint) ?? DefaultVariables.ActiveDirectoryEndpoint;
-
-            clientId = await ExternalVariables.Get(ExternalVariable.AzureSubscriptionClientId, cancellationToken);
-            clientSecret = await ExternalVariables.Get(ExternalVariable.AzureSubscriptionPassword, cancellationToken);
-            tenantId = await ExternalVariables.Get(ExternalVariable.AzureSubscriptionTenantId, cancellationToken);
-            subscriptionId = await ExternalVariables.Get(ExternalVariable.AzureSubscriptionId, cancellationToken);
-
-            resourceGroupName = AzureTestResourceHelpers.GetResourceGroupName();
-
-            resourceGroupLocation = Environment.GetEnvironmentVariable("AZURE_NEW_RESOURCE_REGION") ?? RandomAzureRegion.GetRandomRegionWithExclusions();
-
-            var servicePrincipalAccount = new AzureServicePrincipalAccount(subscriptionId,
-                                                                           clientId,
-                                                                           tenantId,
-                                                                           clientSecret,
-                                                                           "AzureGlobalCloud",
-                                                                           resourceManagementEndpointBaseUri,
-                                                                           activeDirectoryEndpointBaseUri);
-
-            armClient = servicePrincipalAccount.CreateArmClient(retryOptions =>
-                                                                {
-                                                                    retryOptions.MaxRetries = 5;
-                                                                    retryOptions.Mode = RetryMode.Exponential;
-                                                                    retryOptions.Delay = TimeSpan.FromSeconds(2);
-                                                                    retryOptions.NetworkTimeout = TimeSpan.FromSeconds(200);
-                                                                });
-
-            //create the resource group
-            subscriptionResource = armClient.GetSubscriptionResource(SubscriptionResource.CreateResourceIdentifier(subscriptionId));
-
-            await subscriptionResource
-                                 .GetResourceGroups()
-                                 .CreateOrUpdateAsync(WaitUntil.Completed,
-                                                      resourceGroupName,
-                                                      new ResourceGroupData(new AzureLocation(resourceGroupLocation))
-                                                      {
-                                                          Tags =
-                                                          {
-                                                              [AzureTestResourceHelpers.ResourceGroupTags.LifetimeInDaysKey] = AzureTestResourceHelpers.ResourceGroupTags.LifetimeInDaysValue,
-                                                              [AzureTestResourceHelpers.ResourceGroupTags.SourceKey] = AzureTestResourceHelpers.ResourceGroupTags.SourceValue
-                                                          }
-                                                      });
-        }
-
-        [OneTimeTearDown]
-        public async Task Cleanup()
-        {
-            await armClient.GetResourceGroupResource(ResourceGroupResource.CreateResourceIdentifier(subscriptionId, resourceGroupName))
-                           .DeleteAsync(WaitUntil.Started);
-        }
-
-        [Test]
-        [RequiresWindowsServer2016OrAbove("This test requires the az cli, which relies on python 3.10, which doesn't run on windows 2012/2012R2")]
-        public async Task DeployAzureBicepTemplate_PackageSource()
-        {
-            await CommandTestBuilder.CreateAsync<DeployAzureBicepTemplateCommand, Program>()
-                                    .WithArrange(context =>
-                                                 {
-                                                     AddDefaults(context);
-                                                     context.Variables.Add(SpecialVariables.Action.Azure.TemplateSource, "Package");
-                                                     context.Variables.Add(SpecialVariables.Action.Azure.BicepTemplate, "azure_website_template.bicep");
-                                                     context.WithFilesToCopy(packagePath);
-                                                 })
-                                    .Execute();
-        }
-
-        void AddDefaults(CommandTestBuilderContext context)
-        {
-            context.Variables.Add(AzureScripting.SpecialVariables.Account.AccountType, "AzureServicePrincipal");
-            context.Variables.Add(AzureAccountVariables.SubscriptionId, subscriptionId);
-            context.Variables.Add(AzureAccountVariables.TenantId, tenantId);
-            context.Variables.Add(AzureAccountVariables.ClientId, clientId);
-            context.Variables.Add(AzureAccountVariables.Password, clientSecret);
-            context.Variables.Add(SpecialVariables.Action.Azure.ResourceGroupName, resourceGroupName);
-            context.Variables.Add(SpecialVariables.Action.Azure.ResourceGroupLocation, resourceGroupLocation);
-            context.Variables.Add(SpecialVariables.Action.Azure.ResourceGroupDeploymentMode, "Complete");
-            context.Variables.Add(SpecialVariables.Action.Azure.BicepTemplateParameters, ParameterContent);
-
-            context.Variables.Add("SKU", "Standard_LRS");
-            context.Variables.Add("Location", resourceGroupLocation);
-            //storage accounts can be 24 chars long
-            context.Variables.Add("StorageAccountName", AzureTestResourceHelpers.RandomName(length: 24));
-        }
-    }
-}
